SpendingData Object, Location Updating/Editng and Hierarchy Editing (#9)
* Many changes

* Final changes

* Fixing editing image saving

+++ pages/4_Recording_Expenses.py
@@ -3,8 +3,10 @@ import pandas as pd
 import streamlit as st
 from streamlit.delta_generator import DeltaGenerator
 
+import spending
 import utils
 from receipt import Receipt
+from spending import SpendingData
 
 
 def add_item():
@@ -16,7 +18,7 @@ def add_item():
 
 def refresh_all_the_data():
     utils.fetch_transaction_data.clear()
-    utils.fetch_spending_data.clear()
+    spending.SpendingData().fetch_spending_data.clear()
 
 
 def get_info(spending_data: pd.DataFrame):
@@ -96,7 +98,7 @@ def get_info(spending_data: pd.DataFrame):
 
 
 def save_edited_values(edited_df: pd.DataFrame):
-    utils.save_data(edited_df, utils.SPENDING_PATH, utils.SPENDING_SHEET_NAME)
+    utils.save_data(edited_df, spending.SPENDING_PATH, spending.SPENDING_SHEET_NAME)
     refresh_all_the_data()
 
 
@@ -111,7 +113,7 @@ def add_expenses(categorised_transactions: pd.DataFrame):
             on_select="rerun",
             selection_mode="single-row"
         )
-    receipt = upload_display_image()
+    upload_display_image()
     # Form fields
     if prior_expenses_entry.selection.rows:
         new_row = handle_selection_and_prefill(prior_expenses_entry, sorted_df)
@@ -120,23 +122,23 @@ def add_expenses(categorised_transactions: pd.DataFrame):
 
     if st.button("Submit"):
         if st.session_state.receipt:
-            new_row['Receipt Ref'] = receipt.save_image(new_row['Date'], new_row['Receipt Ref'])
+            new_row['Receipt Ref'] = st.session_state.receipt.save_image(new_row['Date'], new_row['Receipt Ref'])
 
         edited_df = pd.concat(
             [categorised_transactions, pd.DataFrame([new_row])],
             ignore_index=True
-        )[utils.SPENDING_DATA_SCHEMA]
+        )[spending.SPENDING_DATA_SCHEMA]
         save_reset(edited_df)
 
 
-def upload_display_image():
+def upload_display_image() -> None:
     uploaded_file = st.file_uploader(
         "Upload receipt images",
         type=["jpg", "jpeg", "png", "pdf"],
         accept_multiple_files=False
     )
     receipt = Receipt()
-    if "uploaded_file" not in st.session_state:
+    if "receipt" not in st.session_state:
         st.session_state.receipt = None
 
     if st.button("Clear image"):
@@ -158,7 +160,6 @@ def upload_display_image():
     if st.session_state.receipt:
         for img in st.session_state.receipt.read_image():
             st.image(img, use_container_width=True)
-    return receipt
 
 
 @st.dialog("Delete expenses", width="large")
@@ -178,13 +179,14 @@ def delete_expenses(categorised_transactions: pd.DataFrame):
         edited_df = (
             sorted_df
             .drop(index=index_label)
-            [utils.SPENDING_DATA_SCHEMA]
+            [spending.SPENDING_DATA_SCHEMA]
             .sort_values(by=["Date"])
             .reset_index()
         )
         if st.button("Delete Transaction"):
             try:
-                st.write(Receipt().set_path(deleted_entry['Receipt Ref']).delete_image())
+                if deleted_entry['Receipt Ref']:
+                    st.write(Receipt().set_path(deleted_entry['Receipt Ref']).delete_image())
             except FileNotFoundError:
                 st.write("File has already been deleted or doesn't exist")
             except ValueError:
@@ -221,22 +223,24 @@ def edit_expenses(categorised_transactions: pd.DataFrame):
             for img in image.read_image():
                 st.image(img, caption=new_row['Receipt Ref'], use_container_width=True)
         else:
-            image = upload_display_image()
+            upload_display_image()
 
         if st.button("Edit transaction"):
-            if image:
-                new_row['Receipt Ref'] = image.save_image(new_row['Date'], new_row['Receipt Ref'])
+            if st.session_state.receipt:
+                if st.session_state.receipt.data:
+                    new_row['Receipt Ref'] = st.session_state.receipt.save_image(new_row['Date'],
+                                                                                 new_row['Receipt Ref'])
             sorted_df.iloc[prior_expenses_entry.selection.rows[0]] = new_row
-            edited_df = sorted_df[utils.SPENDING_DATA_SCHEMA].sort_values(by=["Date"]).reset_index()
+            edited_df = sorted_df[spending.SPENDING_DATA_SCHEMA].sort_values(by=["Date"]).reset_index()
             save_reset(edited_df)
 
 
 def save_reset(edited_df):
     utils.save_data(
         edited_df,
-        utils.SPENDING_PATH,
-        utils.SPENDING_SHEET_NAME)
-    utils.fetch_spending_data.clear()
+        spending.SPENDING_PATH,
+        spending.SPENDING_SHEET_NAME)
+    spending.SpendingData().fetch_spending_data.clear()
     st.session_state.uploaded_file = None
     st.rerun()
 
@@ -342,8 +346,8 @@ def initialise_sidebar(inputs: DeltaGenerator, categorised_transactions: pd.Data
     return start_date, end_date
 
 
-def render_transaction_input(inputs: DeltaGenerator):
-    categorised_transactions = utils.fetch_spending_data().combine().combined
+def render_transaction_input(inputs: DeltaGenerator, spending_data: SpendingData):
+    categorised_transactions = spending_data.combine().combined
     start_date, end_date = initialise_sidebar(inputs, categorised_transactions)
     # We need to offset the end date to be inclusive as the time is set to midnight
     transactions_data = utils.fetch_transaction_data(start_date, end_date + pd.DateOffset(days=1))
@@ -376,10 +380,10 @@ def render_transaction_input(inputs: DeltaGenerator):
                         ~edited_transactions['Location'].isna()]
                 ],
                 ignore_index=True
-            )[utils.SPENDING_DATA_SCHEMA]],
+            )[spending.SPENDING_DATA_SCHEMA]],
         on_click=save_edited_values)
 
 
 # We need to have some sort of way to
 st.set_page_config(layout="wide")
-render_transaction_input(st)
+render_transaction_input(st, spending.SpendingData().fetch_spending_data())
